Brute-force kill all extra xboxdrv instance when shutting down
Shutting down xbox_driver_alternative.py leaves behind processes that keep
control of the USB ports and don't let another xbox driver start correctly.

This is technically not the right way to do it because it can kill nodes running
in another terminal, but it enables easier restarts of field control.